apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.13
alembic==1.13.3
passlib==1.7.4
bcrypt==4.1.3
python-jose==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
